# [examples] Improve `example/matmul` perf by 2.3x 