--- a/examples/pymatmul.py
+++ b/examples/pymatmul.py
@@ -58,4 +58,6 @@ def benchmark_matmul_numpy(M, N, K):
 
 if __name__ == "__main__":
     print("Throughput of a 128x128 matrix multiplication in Python:")
-    benchmark_matmul_python(128, 128, 128)
+    print(benchmark_matmul_python(128, 128, 128))
+    print("Throughput of a 128x128 matrix multiplication in Python numpy:")
+    print(benchmark_matmul_numpy(128, 128, 128))